Update client example.

+++ lib/utils/status.h
@@ -8,7 +8,7 @@ namespace ces::utils {
 
 struct Status {
   int err;             // 0 - OK, ETIME on timeout or errno reported by system.
-  int64_t cnt_bytes;   // number of bytes written or read
+  size_t cnt_bytes;    // number of bytes written or read
   std::string message; // text error message if errno != 0
 };
